+++ sdk/resourcemanager/containerservice/armcontainerservice/constants.go
@@ -7,7 +7,7 @@ package armcontainerservice
 
 const (
 	moduleName    = "github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/containerservice/armcontainerservice"
-	moduleVersion = "v6.6.0"
+	moduleVersion = "v6.7.0"
 )
 
 // AgentPoolMode - A cluster must have at least one 'System' Agent Pool at all times. For additional information on agent
@@ -951,6 +951,32 @@ func PossibleTypeValues() []Type {
 		TypeLast,
 		TypeSecond,
 		TypeThird,
+	}
+}
+
+// UndrainableNodeBehavior - Defines the behavior for undrainable nodes during upgrade. The most common cause of undrainable
+// nodes is Pod Disruption Budgets (PDBs), but other issues, such as pod termination grace period is
+// exceeding the remaining per-node drain timeout or pod is still being in a running state, can also cause undrainable nodes.
+type UndrainableNodeBehavior string
+
+const (
+	// UndrainableNodeBehaviorCordon - AKS will cordon the blocked nodes and replace them with surge nodes during upgrade. The
+	// blocked nodes will be cordoned and replaced by surge nodes. The blocked nodes will have label 'kubernetes.azure.com/upgrade-status:Quarantined'.
+	// A surge node will be retained for each blocked node. A best-effort attempt will be made to delete all other surge nodes.
+	// If there are enough surge nodes to replace blocked nodes, then the upgrade operation and the managed cluster will be in
+	// failed state. Otherwise, the upgrade operation and the managed cluster will be in canceled state.
+	UndrainableNodeBehaviorCordon UndrainableNodeBehavior = "Cordon"
+	// UndrainableNodeBehaviorSchedule - AKS will mark the blocked nodes schedulable, but the blocked nodes are not upgraded.
+	// A best-effort attempt will be made to delete all surge nodes. The upgrade operation and the managed cluster will be in
+	// failed state if there are any blocked nodes.
+	UndrainableNodeBehaviorSchedule UndrainableNodeBehavior = "Schedule"
+)
+
+// PossibleUndrainableNodeBehaviorValues returns the possible values for the UndrainableNodeBehavior const type.
+func PossibleUndrainableNodeBehaviorValues() []UndrainableNodeBehavior {
+	return []UndrainableNodeBehavior{
+		UndrainableNodeBehaviorCordon,
+		UndrainableNodeBehaviorSchedule,
 	}
 }